Fix building tests.
